Move prompts into their own concern
