+++ services/QuillLMS/spec/models/concerns/translatable_spec.rb
@@ -270,12 +270,8 @@ RSpec.describe Translatable do
     let(:locale) { Translatable::DEFAULT_LOCALE }
 
     it 'returns the expected prompt' do
-      expected = <<~STRING
-        You are going to do a translation from english to es-la using simple words and language at a 5th grade reading level. Use shorter words over longer if possible. The tone should be somewhat casual. Return just the translated text preserving (but not translating) the HTML.
-
-        We are translating the instructions for an English-language grammar activity. The content of the activity itself is not translated.
-      STRING
-      expected += "\nTest prompt\n text to translate: "
+      expected = translatable_object.prompt_start(locale:)
+      expected += "Test prompt\n text to translate: "
       expect(translatable_object.open_ai_prompt(locale:)).to eq(expected)
     end
